Create individual headers for each module
