+++ src/key_stack.c
@@ -1,4 +1,4 @@
-#include "rat.h"
+#include "key_stack.h"
 
 KeyStack *new_key_stack() {
   KeyStack *ks = (KeyStack*)malloc(sizeof(*ks));
